수정

+++ src/main/java/footBall/user/UserRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UserRequest {
 
-    private Long fbUserId;           //	유저 ID
+    private int fbUserId;           //	유저 ID
     private String fbUserEmail;      //	유저 이메일
     private String fbUserPswd;       //	유저 비밀번호
     private String fbUserNickname;   //	유저 닉네임
